renamed:    src/tlib.py -> lib/tlib.py 	modified:   tgame.py
